install udev rule for signet

+++ src/ricing.sh
@@ -121,7 +121,14 @@ sudo pacman -S --noconfirm --needed hplip
 sudo pacman -S --noconfirm --needed system-config-printer
 
 # devices
-cd /tmp
-curl -LO https://nthdimtech.com/downloads/signet-releases/0.9.8/gnu-linux/signet-0.9.8.1
-chmod u+x signet-0.9.8.1
-mv signet-0.9.8.1 ~/.config/Scripts/signet && cd -
+if [ ! -f ~/.config/Scripts/signet ]; then
+  cd /tmp
+  curl -LO https://nthdimtech.com/downloads/signet-releases/0.9.8/gnu-linux/signet-0.9.8.1
+  chmod u+x signet-0.9.8.1
+  mv signet-0.9.8.1 ~/.config/Scripts/signet && cd -
+fi
+if [ ! -f /etc/udev/rules.d/50-signet.rules ]; then
+  cd /tmp
+  curl -LO https://nthdimtech.com/downloads/signet-releases/0.9.8/gnu-linux/50-signet.rules
+  sudo mv 50-signet.rules /etc/udev/rules.d/ && cd -
+fi
